Disentangle AboutPage from checking for updates (prepare for the later possibility of introducing AutoUpdate for MSI installs)
